CartItems missing code inserted/Order detail and order working fine. Kindly please send email.

+++ database/migrations/2020_08_06_073628_create_orders_table.php
@@ -15,7 +15,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_number')->unique();
+            $table->double('order_number')->unique();
             $table->string('shipping_address')->nullable();
             $table->string('sub_urb')->nullable();
             $table->string('state')->nullable();
